Introduce WebAssembly backed Web UI
Introduce Web UI for accessing the linter. To make that possible we
compile our linter, tree-sitter, as well as the tree-sitter C language
for WebAssembly and then invoke it from JavaScript on user provided
source code.
Note that neither tree-sitter nor tree-sitter-c can be compiled to WASM
by default. For tree-sitter we instead use the API compatible
tree-sitter-c2rust to make that work and for tree-sitter-c we currently
rely on a fork that adds the necessary support.