web ui: Add support for linting a remote file
If for nothing but illustration purposes, it can come in handy to
quickly be able to retrieve a remote file from a provided URL and lint
it.
This change adds the necessary logic for doing so. To make that as
accessible as possible, provide an example file with some linter
warnings and use its URL as a reasonable default.