Correctly report matches on top most line in report_terminal()
Our line contextualization logic is broken when a match occurs on the
top most line. In such an event we incorrectly start the match report
with the first character of the match, not that of the very first byte,
causing confusing output.
Fix this problem by adjusting the logic accordingly.