Report lint matches in source code order
With multiple lints now being flagged, it is possible to come across a
report where a lint that matched further down in a file is being
reported before one corresponding to a location earlier in the source
code.
Make sure to properly sort matches according to their position inside
the file being linted to prevent such confusing scenarios from the get
go.