Support disabling of lints via code comments
Support the disabling of individual lints via code comments. These
comments take the form 'bpflint: disable=<lint-name>', where <lint-name>
is the name of the lint. Such directives are recognized in a
hierarchical fashion. That is, when placed on a block of sorts (e.g., a
function definition), the apply recursively to the entirety of the
block.

Closes: #3