Embed lint source code directly in build script
So far our build script embedded the path to the actual lint source code
to be include!ed when the generated lint.rs module is built. That is
problematic, because at a conceptual level, it appears as if the build
script can be built on a different machine than it is executed.
To be resilient to this kind of setup, just read and embed the lint
source code directly into the generated lint.rs module.