Introduce concept of "lint internal captures"
Introduce the concept of "internal captures" for use in lints. An
internal capture is one whose begins with a double underscore and it is
considered internal to the lint and will not be reported by the linting
infrastructure. This can be useful to enforce certain constraints on
matches, e.g., by checking that an identifier matches a name without
triggering a lint message. In tree-sitter's query language it is not
otherwise possible to check such constraints without creating a named
capture.